clear command

+++ commands/clear.js
@@ -1,18 +1,19 @@
 module.exports = {
     name: 'c',
-    description: 'deleted specified number of messages',
+    description: 'supprime TOUT, genre une BOMBE BOUMMMM',
     execute(client, api, config, message, args) {
-        var amount = parseInt(args[0]) + 1;
-
-        if (isNaN(amount)) {
-            amount = 100
-        } else if (amount < 1 || amount > 99) {
-            amount = 100
+        message.delete()
+        if (message.member.roles.cache.find(r => r.name == "Admin")) {
+            var amount = parseInt(args[0]) + 1;
+            if (isNaN(amount)) {
+                amount = 100
+            } else if (amount < 1 || amount > 99) {
+                amount = 100
+            }
+            message.channel.bulkDelete(amount, true).catch(err => {
+                console.log(err);
+                message.channel.send("Wola deso je peux là j'ai poney");
+            });
         }
-
-        message.channel.bulkDelete(amount, true).catch(err => {
-            console.log(err);
-            message.channel.send("Wola deso je peux la j'ai poney");
-        });
     },
 };
